% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_ripleys_k2.R
\name{bi_ripleys_k2}
\alias{bi_ripleys_k2}
\title{Bivariate Ripley's K}
\usage{
bi_ripleys_k2(
  mif,
  mnames,
  r_range = 0:100,
  edge_correction = "translation",
  num_permutations = 50,
  permute = FALSE,
  keep_permutation_distribution = FALSE,
  overwrite = TRUE,
  workers = 6,
  big = 1000,
  nlarge = 1000
)
}
\arguments{
\item{mnames}{vector of column names for phenotypes or data frame of marker combinations}

\item{r_range}{vector range of radii to calculate co-localization *K*}

\item{edge_correction}{character edge_correction method, one of "translation", "border", "or none"}

\item{num_permutations}{integer number of permutations to estimate CSR}

\item{keep_permutation_distribution}{boolean as to whether to summarise permutations to mean}

\item{overwrite}{boolean as to whether to replace existing bivariate_Count if exists}

\item{workers}{integer number of CPU workers to use when number of `anchor` or `counted` is greater than 10,000}

\item{big}{integer used as the threshold for subsetting large samples, default is 1000 either *i* or *j*}

\item{nlarge}{number of cells in either *i* or *j* to flip to no edge correction - at small (relative to whole spatial region) *r* values differences in results between correction methods is negligible so running a few samples is recommended. Perhaps compute outweighs small differences in correction methods.}

\item{spatial}{spatial data frame with columns of XMin, XMax, YMin, YMax}
}
\value{
mif object with bivariate ripley's K calculated
}
\description{
Bivariate Ripley's K
}
\examples{
x <- spatialTIME::create_mif(clinical_data = spatialTIME::example_clinical \%>\% 
                               dplyr::mutate(deidentified_id = as.character(deidentified_id)),
                             sample_data = spatialTIME::example_summary \%>\% 
                               dplyr::mutate(deidentified_id = as.character(deidentified_id)),
                             spatial_list = spatialTIME::example_spatial,
                             patient_id = "deidentified_id", 
                             sample_id = "deidentified_sample")
mnames_good <- c("CD3..Opal.570..Positive","CD8..Opal.520..Positive",
                 "FOXP3..Opal.620..Positive","PDL1..Opal.540..Positive",
                 "PD1..Opal.650..Positive","CD3..CD8.","CD3..FOXP3.")
x2 = bi_ripleys_k2(mif = x, mnames = mnames_good, 
                   r_range = 0:100, edge_correction = "translation", 
                   num_permutations = 50, keep_permutation_distribution = FALSE, 
                   workers = 6, big = 1000)
}
