% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripleys_k.R
\name{bi_ripleys_k_v2}
\alias{bi_ripleys_k_v2}
\title{Calculate Bivariate Count Based Measures of Spatial Clustering function for IF data}
\usage{
bi_ripleys_k_v2(
  mif,
  mnames,
  r_range = seq(0, 100, 50),
  num_permutations = 50,
  edge_correction = "translation",
  method = "K",
  keep_perm_dis = FALSE,
  exhaustive = TRUE
)
}
\arguments{
\item{mif}{An MIF object}

\item{mnames}{Character vector of marker names to estimate degree of
spatial clustering. Spatial clustering will be computed between each
combination of markers in this list.}

\item{r_range}{Numeric vector of potential r values this range must include 0}

\item{num_permutations}{Numeric value indicating the number of permutations used.
Default is 50.}

\item{edge_correction}{Character value indicating the type of edge correction
to use. Options include "theoretical", "translation", "isotropic" or "border".
Various edges corrections are most appropriate in different settings. Default
is "translation".}

\item{method}{Character value indicating which measure (K, L, M) used to
estimate the degree of spatial clustering. Description of the methods can be
found in Details section.}

\item{keep_perm_dis}{Logical value determining whether or not to keep the full
distribution of permuted K values}

\item{exhaustive}{Logical. If TRUE then markers must be a vector and spatial
measures will be computed all pairs of unique markers. If FALSE then markers must
be a data.frame with the desired combinations.}

\item{id}{Character string of variable name for subject ID in TMA data.}
}
\value{
Returns a data frame
\item{anchor}{Marker for which the distances are measured from}
\item{counted}{Marker for which the distances are measured to}
\item{Theoretical CSR}{Expected value assuming complete spatial randomness}
\item{Permuted CSR}{Average observed K, L, or M for the permuted point
process}
\item{Observed}{Observed value for the observed point process}
\item{Degree of Clustering Permuted}{Degree of spatial clustering where the
reference is the permutated estimate of CSR}
\item{Degree of Clustering Theoretical}{Degree of spatial clustering where the
reference is the theoretical estimate of CSR}
}
\description{
This function calculates count based Measures (Ripley's K, Besag
L, and Marcon's M) of IF data to characterize correlation of spatial point
process for two markers.
}
