% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_perms}
\alias{compute_perms}
\title{Compute the permutations of permuations of a distance matrix}
\usage{
compute_perms(perms1, r_val, distances1, edge1, area)
}
\arguments{
\item{r_val}{a single r value which to count within for Ripley's K}

\item{distances1, edge1}{matrices containing equal columns and rows for distances between cells and edge correction corresponding to cell pairs}

\item{area}{a single value that contains the area of the window around points}

\item{perm1}{a numeric matrix containing rows in `distances1` and `edge1` to use as positive cell locations}
}
\value{
a vector of K statistics of length nrow(perms1)
}
\description{
Compute the permutations of permuations of a distance matrix
}
\examples{
perms1 = as.matrix(sample(1:10, 5, replace = F))
r_val = 1
distances1 = matrix(nrow = 10, data = abs(rnorm(100)))
edge1 = distances1 #no edge correction
area = 100
K = compute_perms(perms1, r_val, distances1, edge1, area)
}
