% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordmerge.R
\name{CoordMerge}
\alias{CoordMerge}
\title{CoordMerge
@param files Seurat object as Input
@param pattern Cluster names in Idents}
\usage{
CoordMerge(files = "", pattern = NULL)
}
\description{
@details
This function calculates and adds coordinates values to each line drawn in data frame.
}
\details{
@import Seurat
@import tidyverse
@import fs
@export
}
