% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordmerge.R
\name{CoordMergeHD}
\alias{CoordMergeHD}
\title{CoordMergeHD}
\usage{
CoordMergeHD(files = "", pattern = "")
}
\arguments{
\item{files}{Seurat object as Input}

\item{pattern}{Cluster names in Idents}
}
\description{
CoordMergeHD
}
\details{
This function calculates and adds coordinates values to each line drawn in data frame for Spatial HD.
}
