% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-example-tma.R
\docType{data}
\name{example_tma}
\alias{example_tma}
\title{Example list of spatial TMA data}
\format{
A list of 10 data frames:
\itemize{
\item TMA_\link{T,3}.tiff
\item TMA_\link{I,7}.tiff
\item TMA_\link{W,5}.tiff
\item TMA_\link{N,1}.tiff
\item TMA_\link{D,2}.tiff
\item TMA_\link{C,6}.tiff
\item TMA_\link{F,6}.tiff
\item TMA_\link{F,3}.tiff
\item TMA_\link{P,5}.tiff
\item TMA_\link{T,9}.tiff
}
}
\usage{
example_tma
}
\description{
A list containing 10 spatial data frames
}
\keyword{datasets}
