% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseudoTime.R
\name{GeneGet}
\alias{GeneGet}
\title{GeneGet}
\usage{
GeneGet(var = NULL, n_clusters = 2)
}
\arguments{
\item{var}{Pheatmap file from Pseudo2Time}

\item{n_clusters}{Number of clusters from heatmap}
}
\description{
GeneGet
}
\details{
Get genes present in each clusters identified from heatmap.
}
