% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialVis.R
\name{GeneVis}
\alias{GeneVis}
\title{GeneVis}
\usage{
GeneVis(data = NULL, genes = NULL)
}
\arguments{
\item{data}{Seurat object with pseudotime values in metadata}

\item{genes}{Genes to be plotted}
}
\description{
GeneVis
}
\details{
Visualization of genes of interest using reference line as starting point
}
