% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseudoTime.R
\name{PseudoM3Time}
\alias{PseudoM3Time}
\title{PseudoM3Time}
\usage{
PseudoM3Time(
  file = NULL,
  assay = c("RNA", "SCT"),
  values = c("pt", "st"),
  q_cutoff = 0.01,
  morans_cutoff = 0.05,
  cores = 4,
  return_obj = T
)
}
\arguments{
\item{file}{Seurat object}

\item{assay}{Assays to choose}

\item{values}{Use pseudotime or spatialtime}

\item{q_cutoff}{q value cutoff}

\item{morans_cutoff}{morans values cutoff}

\item{cores}{CPU cores}

\item{return_obj}{Return full object}
}
\description{
PseudoM3Time
}
\details{
This function calculates pseudotime using Monocle3
}
