% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseudoTime.R
\name{PseudoTime}
\alias{PseudoTime}
\title{PseudoTime}
\usage{
PseudoTime(
  file = NULL,
  assay = "RNA",
  min_expr = 0.1,
  min_cells = 3,
  mean_expr = 0.1,
  pvalue = 0.05,
  cores = 1,
  return_obj = F
)
}
\arguments{
\item{file}{Seurat object with module annotation}

\item{assay}{Seurat Assay}

\item{min_expr}{Minimum gene expression}

\item{min_cells}{minimum cells expression}

\item{mean_expr}{Mean gene expression}

\item{pvalue}{pvalue threashold to filter genes out}

\item{cores}{number of CPU cores to use}

\item{return_obj}{plot monocle heatmap}
}
\description{
PseudoTime
}
\details{
This function calculates and adds coordinates values to each line drawn in data frame.
}
