% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripleys_k2.R
\name{ripleys_k2}
\alias{ripleys_k2}
\title{Calculate Ripley's K}
\usage{
ripleys_k2(
  mif,
  mnames,
  r_range = seq(0, 100, 1),
  num_permutations = 50,
  edge_correction = "translation",
  method = "K",
  keep_perm_dis = FALSE,
  workers = 1,
  overwrite = F,
  xloc = NULL,
  yloc = NULL
)
}
\arguments{
\item{mif}{object of class \code{mif} created with \code{create_mif}}

\item{mnames}{cell phenotype markers to calculate Ripley's K for}

\item{r_range}{radius range (including 0)}

\item{num_permutations}{number of permutations to use to estimate CSR. If \code{keep_perm_dis} is set to FALSE, this will be ignored}

\item{edge_correction}{edge correction method to pass to \code{Kest}. can take one of "translation", "isotropic", "none"}

\item{method}{not used currently}

\item{keep_perm_dis}{whether to keep the permutation results. Setting to FALSE will cause skipping of permutations and use CSR estimate}

\item{workers}{number of cores to use for calculations}

\item{overwrite}{whether to overwrite the \code{univariate_Count} slot within \code{mif$derived}}

\item{xloc, }{yloc the location of the center of cells. If left \code{NULL}, \code{XMin}, \code{XMax}, \code{YMin}, and \code{YMax} must be present.}
}
\value{
object of class \code{mif}
}
\description{
Calculate Ripley's K
}
\examples{
x <- spatialTIME::create_mif(clinical_data =spatialTIME::example_clinical \%>\% 
  dplyr::mutate(deidentified_id = as.character(deidentified_id)),
  sample_data = spatialTIME::example_summary \%>\% 
  dplyr::mutate(deidentified_id = as.character(deidentified_id)),
  spatial_list = spatialTIME::example_spatial,
  patient_id = "deidentified_id", 
  sample_id = "deidentified_sample")
mnames = x$spatial[[1]] \%>\%
  colnames() \%>\%
  grep("Pos|CD", ., value = T) \%>\%
  grep("Cyto|Nucle", ., value = T, invert = T)
x2 = ripleys_k2(mif = x, 
  mnames = mnames, 
   r_range = seq(0, 100, 1), 
    num_permutations = 100,
     edge_correction = "translation", 
      method = "K", 
      keep_perm_dis = F, 
      workers = 1, 
       overwrite = T)
}
