% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripleys_k_auc.R
\name{ripleys_k_auc}
\alias{ripleys_k_auc}
\title{Calculate Area under Ripley's K curve}
\usage{
ripleys_k_auc(mif)
}
\arguments{
\item{mif}{object of class `mif` made with `createMIF`}
}
\value{
object of class `mif` with `univariate_AUC` in derived slot
}
\description{
Calculate Area under Ripley's K curve
}
\examples{
x <- spatialTIME::create_mif(clinical_data =spatialTIME::example_clinical \%>\% 
  dplyr::mutate(deidentified_id = as.character(deidentified_id)),
  sample_data = spatialTIME::example_summary \%>\% 
  dplyr::mutate(deidentified_id = as.character(deidentified_id)),
  spatial_list = spatialTIME::example_spatial,
  patient_id = "deidentified_id", 
  sample_id = "deidentified_sample")
mnames = "CD3..Opal.570..Positive"
x2 = ripleys_k2(mif = x,
  mnames = mnames, 
  r_range = seq(0, 100, 1), 
  num_permutations = 100,
  edge_correction = "translation", 
  method = "K", 
  keep_perm_dis =FALSE, 
  workers = 1, 
  overwrite =TRUE)
x2 = ripleys_k_auc(x2)

}
