% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialVis.R
\name{ScaleRatio}
\alias{ScaleRatio}
\title{ScaleRatio}
\usage{
ScaleRatio(object = NULL)
}
\arguments{
\item{object}{Seurat object input}
}
\description{
ScaleRatio
}
\details{
Image scale adjustment for SpatialDimPlot
}
