% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shinyapp.R
\name{ShinySelection}
\alias{ShinySelection}
\title{ShinyApp}
\usage{
ShinySelection(file = NULL, file.name = NULL)
}
\arguments{
\item{file}{Seurat object}

\item{file.name}{Coordinates spots output file}
}
\description{
ShinyApp
}
\details{
Shiny app for manual spots selection and coordinates export.
}
