% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shinyapp.R
\name{ShinySpots}
\alias{ShinySpots}
\title{ShinySpots}
\usage{
ShinySpots(seurat_obj = NULL, coord_file = NULL, slice.n = "slice1")
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{coord_file}{Spots selection coordinates}

\item{slice.n}{Slice number}
}
\description{
ShinySpots
}
\details{
This function allows coordinates extraction from specific spots barcodes previously selected in ShinySelection function
}
