% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mif.R
\name{simulate_mif}
\alias{simulate_mif}
\title{Simulate Multiplex Immunoflourescent data}
\usage{
simulate_mif(
  k = 10,
  xmin = 0,
  xmax = 10,
  ymin = 0,
  ymax = 10,
  sdmin = 0.5,
  sdmax = 2,
  k_cell_stroma = 3,
  k_cell_tumor = 5,
  upper_lim_tumor = 0.1
)
}
\arguments{
\item{k}{numeric value ...}

\item{xmin}{numeric value ...}

\item{xmax}{numeric value ...}

\item{ymin}{numeric value ...}

\item{ymax}{numeric value ...}

\item{sdmin}{numeric value ...}

\item{sdmax}{numeric value ...}

\item{k_cell_stroma}{numeric value ...}

\item{k_cell_tumor}{numeric value ...}

\item{upper_lim_tumor}{numeric value ...}
}
\description{
Simulate IF data
}
