% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial2time.R
\name{Spatial2Time}
\alias{Spatial2Time}
\title{Spatial2Time}
\usage{
Spatial2Time(file = NULL, file2 = NULL, id = NULL)
}
\arguments{
\item{file}{Seurat object as Input}

\item{id}{Export coordinates files}

\item{fil2}{Cluster names in Idents}
}
\description{
Spatial2Time
}
\details{
This function calculates and adds coordinates values to each line drawn in data frame.
}
