% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial2time.R
\name{Spatial2TimeHD}
\alias{Spatial2TimeHD}
\title{Spatial2TimeHD}
\usage{
Spatial2TimeHD(file = NULL, file2 = NULL, id = NULL)
}
\arguments{
\item{file}{Seurat's cluster coordinates as input}

\item{file2}{Seurat's reference cluster coordinates as input}

\item{id}{Export coordinates files}
}
\description{
Spatial2TimeHD
}
\details{
This function calculates and adds coordinates values to each line drawn in data frame.
}
