% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCalc.R
\name{SpatialCalc}
\alias{SpatialCalc}
\title{SpatialCalc}
\usage{
SpatialCalc(file = "", factor = 1, colors = NULL, tissue = NULL)
}
\arguments{
\item{file}{Fiji file containing drawed pixel coordinates}

\item{factor}{Seurat image scale factor}

\item{colors}{Fiji line colors}

\item{tissue}{Cluster identities assigned}
}
\description{
SpatialCalc
}
\details{
Scale exported clusters coordinates to their respective drawed line colors using image factor
}
