% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial2time.R
\name{SpatialShinyTime}
\alias{SpatialShinyTime}
\title{SpatialShinyTime}
\usage{
SpatialShinyTime(file = NULL, file2 = NULL, id = NULL)
}
\arguments{
\item{file}{Seurat's cluster coordinates as input}

\item{file2}{Seurat coordinates of reference spots selected in ShinyApp}

\item{id}{Don't remember this}
}
\description{
SpatialShinyTime
}
\details{
This function calculates euclidean distances of reference spots selected in ShinyApp against cluster coordinates of interest
}
