% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialTime.R
\name{SpatialTime}
\alias{SpatialTime}
\title{SpatialTime}
\usage{
SpatialTime(file = NULL, reference = NULL, compare = NULL)
}
\arguments{
\item{file}{file containing clusters coordinates scaled by image factor}

\item{reference}{cluster as reference}

\item{compare}{clusters spots to compare against reference}
}
\description{
SpatialTime
}
\details{
This function calculates and adds coordinates values to each line drawn in data frame.
}
