% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialVis.R
\name{SpatialVis}
\alias{SpatialVis}
\title{SpatialVis}
\usage{
SpatialVis(
  file = NULL,
  st.calc = NULL,
  spatial.by = c("abs", "rel"),
  slice = "slice1",
  remove.na = F,
  return_obj = T,
  pt_size = 1,
  image_opacity = 1
)
}
\arguments{
\item{file}{Seurat Object as Input}

\item{st.calc}{Spatialtime values}

\item{spatial.by}{absolute or relative values}

\item{slice}{Select tissue slice}

\item{remove.na}{subset only tissue of spatialtime}

\item{return_obj}{Return object}

\item{pt_size}{change spatial spots size}

\item{image_opacity}{H&E background opacity

SpatialVis}
}
\description{
SpatialVis
}
\details{
This function calculates and adds coordinates values to each line drawn in data frame.
}
