% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset2Labels.R
\name{subset2Labels}
\alias{subset2Labels}
\title{subset2Labels}
\usage{
subset2Labels(data = NULL, cluster = NULL, export.all = TRUE, dir.out = "")
}
\arguments{
\item{data}{Seurat object as Input}

\item{cluster}{Cluster names in Idents}

\item{export.all}{Export coordinates files}

\item{dir.out}{directory output}
}
\description{
subset2Labels
}
\details{
Export clusters coordinates and all their resptive slices
}
