% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset2Labels.R
\name{subset2LabelsHD}
\alias{subset2LabelsHD}
\title{subset2LabelsHD}
\usage{
subset2LabelsHD(
  data = NULL,
  cluster = NULL,
  export.all = FALSE,
  dir.out = "coord_out"
)
}
\arguments{
\item{data}{Seurat object as Input}

\item{cluster}{Cluster names in Idents}

\item{export.all}{Export coordinates files}

\item{dir.out}{directory output}
}
\description{
subset2LabelsHD
}
\details{
Export clusters coordinates and all their resptive slices
}
