% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetLabels.R
\name{subsetLabels}
\alias{subsetLabels}
\title{subsetLabels
@param file Seurat object as Input
@param cluster Cluster names in Idents
@param export.all Export coordinates files
@param slice.n Image slice number
@param dir.out directory output}
\usage{
subsetLabels(
  file = NULL,
  cluster = NULL,
  export.all = TRUE,
  slice.n = "slice1",
  dir.out = ""
)
}
\description{
@details
This function calculates and adds coordinates values to each line drawn in data frame.
}
\details{
@import Seurat
@import tidyverse
@export
}
