% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripleys_k.R
\name{bi_ripleys_k}
\alias{bi_ripleys_k}
\title{Bivariate Count Based Measures of Spatial Clustering function for IF data}
\usage{
bi_ripleys_k(
  mif,
  mnames,
  r_range = seq(0, 100, 50),
  num_permutations = 50,
  edge_correction = "translation",
  method = "K",
  keep_perm_dis = FALSE,
  exhaustive = TRUE,
  workers = 1,
  overwrite = FALSE,
  xloc = NULL,
  yloc = NULL
)
}
\arguments{
\item{mif}{An MIF object}

\item{mnames}{Character vector of marker names to estimate degree of
spatial clustering. Spatial clustering will be computed between each
combination of markers in this list.}

\item{r_range}{Numeric vector of potential r values this range must include 0}

\item{num_permutations}{Numeric value indicating the number of permutations used.
Default is 50.}

\item{edge_correction}{Character value indicating the type of edge correction
to use. Options include "theoretical", "translation", "isotropic" or "border".
Various edges corrections are most appropriate in different settings. Default
is "translation".}

\item{method}{Character value indicating which measure (K, L, M) used to
estimate the degree of spatial clustering. Description of the methods can be
found in Details section.}

\item{keep_perm_dis}{Logical value determining whether or not to keep the full
distribution of permuted K values}

\item{exhaustive}{Logical. If TRUE then markers must be a vector and spatial
measures will be computed all pairs of unique markers. If FALSE then markers must
be a data.frame with the desired combinations.}

\item{workers}{Integer value for the number of workers to spawn}

\item{overwrite}{Logical value determining if you want the results to replace the
current output (TRUE) or be to be appended (FALSE).}

\item{xloc}{a string corresponding to the x coordinates. If null the average of
XMin and XMax will be used}

\item{yloc}{a string corresponding to the y coordinates. If null the average of
YMin and YMax will be used}
}
\value{
Returns a data frame
\item{anchor}{Marker for which the distances are measured from}
\item{counted}{Marker for which the distances are measured to}
\item{Theoretical CSR}{Expected value assuming complete spatial randomness}
\item{Permuted CSR}{Average observed K, L, or M for the permuted point
process}
\item{Observed}{Observed value for the observed point process}
\item{Degree of Clustering Permuted}{Degree of spatial clustering where the
reference is the permuted estimate of CSR}
\item{Degree of Clustering Theoretical}{Degree of spatial clustering where the
reference is the theoretical estimate of CSR}
}
\description{
This function calculates count based Measures (Ripley's K, Besag
L, and Marcon's M) of IF data to characterize correlation of the observed and permyted spatial point
processes for two markers.
}
\examples{
#' #Create mif object
library(dplyr)
x <- create_mif(clinical_data = example_clinical \%>\% 
mutate(deidentified_id = as.character(deidentified_id)),
sample_data = example_summary \%>\% 
mutate(deidentified_id = as.character(deidentified_id)),
spatial_list = example_spatial,
patient_id = "deidentified_id", 
sample_id = "deidentified_sample")

#Ripley's K for the colocalization of CD3+CD8+ positive cells and
#CD3+FOXP3+ positive cells where CD3+FOXP3+ is the reference cell type at 
#neighborhood size of 10,20,...,100 (zero must be included in the input).

x <- bi_ripleys_k(mif = x, mnames = c("CD3..CD8.", "CD3..FOXP3."), 
num_permutations = 1, edge_correction = 'translation', r = seq(0,100,10),
keep_perm_dis = FALSE, workers = 1, exhaustive = TRUE) 

}
