% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-MIF.R
\name{create_MIF}
\alias{create_MIF}
\title{Create Multiplex Immunoflourescent object}
\usage{
create_MIF(spatial_list, clinical_data = NULL, sample_data = NULL)
}
\arguments{
\item{clinical_data}{A data frame containing patient level data. Patient ID
indicated with variable \code{patient_id}}

\item{sample_data}{A data frame containing sample level data. Sample ID
should be indicated with variable \code{image.tag} while patient ID should
be indicated with variable  \code{patient_id}.}

\item{spatial_lsit}{A named list of data frames with the spatial data from
each sample making up each individual data frame}
}
\value{
Returns a custom MIF
\item{spatial}{Named list of spatial data}
\item{clinical}{Data frame of clinical data}
\item{sample}{Data frame of sample data}
\item{derived}{List of data derived using the MIF object}
}
\description{
Creates an MIF object for use in spatialIF functions
}
