% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mif.R
\name{create_mif}
\alias{create_mif}
\title{Create Multiplex Immunoflourescent object}
\usage{
create_mif(
  spatial_list,
  clinical_data = NULL,
  sample_data = NULL,
  clean_columns = TRUE
)
}
\arguments{
\item{spatial_list}{A named list of data frames with the spatial data from
each sample making up each individual data frame}

\item{clinical_data}{A data frame containing patient level data. Patient ID
indicated with variable \code{patient_id}}

\item{sample_data}{A data frame containing sample level data. Sample ID
should be indicated with variable \code{image.tag} while patient ID should
be indicated with variable  \code{patient_id}.}

\item{clean_columns}{A logical value indicating if names in spatial data frames
should be rewritten in a cleaner format. Default is TRUE.}
}
\value{
Returns a custom MIF
\item{spatial}{Named list of spatial data}
\item{clinical}{Data frame of clinical data}
\item{sample}{Data frame of sample data}
\item{derived}{List of data derived using the MIF object}
}
\description{
Creates an MIF object for use in spatialIF functions
}
