% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-example-clinical.R
\docType{data}
\name{example_clinical}
\alias{example_clinical}
\title{Clinical variables of 229 patients}
\format{
A tibble with 229 rows and 6 variables
\describe{
\item{age}{age at diagnosis}
\item{race}{self-idenitifed race}
\item{sex}{patient biological sex}
\item{status}{disease status}
\item{deidenitifed_sample}{sample identifier}
\item{deidentified_id}{patient identifier}
}
}
\usage{
example_clinical
}
\description{
A tibble wuith clinical characteristics for 229 patients
}
\keyword{datasets}
