% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-example-spatial.R
\docType{data}
\name{example_spatial}
\alias{example_spatial}
\title{Example list of 5 spatial TMA data}
\format{
A list of 5 data frames: 
\itemize{
\item TMA_\[3,B\].tiff
\item TMA_\[6,F\].tiff
\item TMA_\[7,B\].tiff
\item TMA_\[9,K\].tiff
\item TMA_\[8,U\].tiff
}
}
\usage{
example_spatial
}
\description{
A list containing 5 spatial data frames
}
\keyword{datasets}
