% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-example-summary.R
\docType{data}
\name{example_summary}
\alias{example_summary}
\title{Marker summaries of 229 samples}
\format{
A tibble with 229 rows and 29 variables: 
\describe{
  \item{deidentified_id}{patient-level id}
  \item{deidentified_sample}{sample-level id}
  ...
}
}
\usage{
example_summary
}
\description{
A dataset containing summaries of 25 markers and 229 samples
}
\keyword{datasets}
