% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-example-tma.R
\docType{data}
\name{example_tma}
\alias{example_tma}
\title{Example list of spatial TMA data}
\format{
A list of 10 data frames:
\itemize{
\item TMA_[T,3].tiff
\item TMA_[I,7].tiff
\item TMA_[W,5].tiff
\item TMA_[N,1].tiff
\item TMA_[D,2].tiff
\item TMA_[C,6].tiff
\item TMA_[F,6].tiff
\item TMA_[F,3].tiff
\item TMA_[P,5].tiff
\item TMA_[T,9].tiff
}
}
\usage{
example_tma
}
\description{
A list containing 10 spatial data frames
}
\keyword{datasets}
