% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialVis.R
\name{GeneVis}
\alias{GeneVis}
\title{GeneVis}
\usage{
GeneVis(
  file = NULL,
  column = NULL,
  signal = c("gene", "pathway"),
  span = 1,
  se = F,
  line_thickness = 1
)
}
\arguments{
\item{file}{Seurat object with pseudotime values in metadata}

\item{column}{Genes to be plotted}

\item{signal}{Selecting wheter genes or module pathways to visualize}

\item{span}{Curve smoothness}

\item{se}{Standard error}

\item{line_thickness}{Curve plot line thickness}
}
\description{
GeneVis
}
\details{
Visualization of genes of interest using reference line as starting point
}
