% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialVis.R
\name{GeneVis}
\alias{GeneVis}
\title{GeneVis}
\usage{
GeneVis(file = NULL, column = NULL, signal = c("gene", "pathway"))
}
\arguments{
\item{file}{Seurat object with pseudotime values in metadata}

\item{column}{Genes to be plotted}

\item{signal}{Selecting wheter genes or module pathways to visualize}
}
\description{
GeneVis
}
\details{
Visualization of genes of interest using reference line as starting point
}
