% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_immunoflo.R
\name{plot_immunoflo}
\alias{plot_immunoflo}
\title{Generate plot of TMA point process}
\usage{
plot_immunoflo(
  mif,
  plot_title,
  mnames,
  mlabels = NULL,
  mcolors = NULL,
  dark_mode = FALSE,
  cell_type = NULL,
  filename = NULL,
  path = NULL
)
}
\arguments{
\item{mif}{MIF object created using create_MIF().}

\item{plot_title}{Character string or vector of character strings of variable name(s) to serve as plot title(s).}

\item{mnames}{Character vector containing marker names.}

\item{mlabels}{Character vector of label for marker names to display in plot legend.}

\item{mcolors}{Character vector of color names to display markers in the plot.}

\item{dark_mode}{Plot using dark color scheme}

\item{cell_type}{Character vector of cell type ???}

\item{filename}{Character string of file name to store plots. Plots are generated as single .pdf file.}

\item{path}{Different path than file name or to use in conjunction with filename ???}
}
\value{
A list of ggplots (the length of spatial data frames) - one for each TMA or ROI
}
\description{
This function generates plot of point process in rectangular or circular window.
}
