% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseudoTime.R
\name{PseudoTime}
\alias{PseudoTime}
\title{PseudoTime}
\usage{
PseudoTime(
  file = NULL,
  assay = "RNA",
  min_expr = 0.1,
  min_cells = 2,
  mean_expr = 0.1
)
}
\arguments{
\item{file}{Seurat object with module annotation}

\item{assay}{Seurat Assay}

\item{min_expr}{Minimum gene expression}

\item{min_cells}{minimum cells expression}

\item{mean_expr}{Mean gene expression}
}
\description{
PseudoTime
}
\details{
This function calculates and adds coordinates values to each line drawn in data frame.
}
