% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripleys_k.R
\name{ripleys_k}
\alias{ripleys_k}
\title{Calculate Ripley's K function for IF data}
\usage{
ripleys_k(
  mif,
  id,
  mnames,
  wshape = "irregular",
  r_range = seq(0, 100, 50),
  dist = 200,
  kestimation = FALSE
)
}
\arguments{
\item{mif}{An MIF object}

\item{id}{Character string of variable name for subject ID in TMA data.}

\item{mnames}{Character vector of marker names to calculate Ripley's K on.}

\item{wshape}{Character string of window shape. Potenital values are
"r" or "rectangle" for rectanglular window, "c" or "circle" for
circular window and "i" or "irregular" for a custom polygon.}

\item{r_range}{Numeric vector of potential r values to estimate K at.}

\item{dist}{Numeric value for something (Brooke or Chris please fill in)}

\item{kestimation}{Logical value determining the type estimation performed.
TRUE estimates Ripley's reduced second moment function while FALSE
estimates Besags's transformation of Ripley's K}
}
\value{
Returns a data frame
\item{r}{Subject ID in TMA data}
\item{theo}{Ripley's K estimate using translation edge correction}
\item{border}{Ripley's K estimate using isotropic edge correction}
\item{trans}{Theoritical value of Ripley's K}
\item{iso}{Intensity of TMA data}
}
\description{
This function calculates Ripley's K function of IF data to
characterize correlation of spatial point process using tranlation and
isotropic edge correction method.
}
