% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shinyapp.R
\name{ShinySelection}
\alias{ShinySelection}
\title{ShinySelection}
\usage{
ShinySelection(
  file = NULL,
  id = NULL,
  file.name = "selected_spots_random.csv",
  angle = -90
)
}
\arguments{
\item{file}{Seurat object}

\item{id}{Seurat object metadata column with cluster labels}

\item{file.name}{Coordinates spots output file}

\item{angle}{Plotly SpatialFeaturePlot rotation angle}
}
\description{
ShinySelection
}
\details{
Shiny app for manual spots selection and coordinates export.
}
