% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shinyapp.R
\name{ShinySelection}
\alias{ShinySelection}
\title{ShinySelection}
\usage{
ShinySelection(file = NULL, id = NULL, file.name = "selected_spots.csv")
}
\arguments{
\item{file}{Seurat object}

\item{file.name}{Coordinates spots output file}
}
\description{
ShinySelection
}
\details{
Shiny app for manual spots selection and coordinates export.
}
