% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial2time.R
\name{Spatial2Time}
\alias{Spatial2Time}
\title{Spatial2Time
@param file Seurat object as Input
@param fil2 Cluster names in Idents
@param id Export coordinates files}
\usage{
Spatial2Time(file = NULL, file2 = NULL, id = NULL)
}
\description{
@details
This function calculates and adds coordinates values to each line drawn in data frame.
}
\details{
@import Seurat
@import tidyverse
@export
}
