% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCalc.R
\name{SpatialCalc}
\alias{SpatialCalc}
\title{SpatialCalc}
\usage{
SpatialCalc(file = "", factor = 1, colors = NULL, tissue = NULL)
}
\arguments{
\item{file}{Fiji file}

\item{factor}{Scale factor}

\item{colors}{Line colors}

\item{tissue}{Tissue ID}
}
\description{
SpatialCalc
}
\details{
This function calculates and adds coordinates values to each line drawn in data frame.
}
