% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialTime.R
\name{SpatialTime}
\alias{SpatialTime}
\title{SpatialTime}
\usage{
SpatialTime(file = "", id = NULL)
}
\arguments{
\item{file}{cluster coordinates file}

\item{id}{line as reference

@details
This function calculates and adds coordinates values to each line drawn in data frame.}
}
\description{
SpatialTime
}
