% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_mif.R
\name{subset_mif}
\alias{subset_mif}
\title{Subset mif object on cellular level}
\usage{
subset_mif(mif, classifier, level, markers)
}
\arguments{
\item{mif}{An MIF object}

\item{classifier}{Column name for spatial dataframe to subset}

\item{level}{Determines which level of the classifier to keep.}

\item{markers}{vector of}
}
\value{
mif object where the spatial list only as the cell that are the specified level.
}
\description{
This function allows to subset the mif object into compartments. 
For instance a mif object includes all cells and the desired analysis is based
on only the tumor or stroma compartment then this function will subset the 
spatial list to just the cells in the desired compartment
}
\examples{
#Create mif object
library(dplyr)
mif <- create_mif(
 clinical_data = example_clinical,
  sample_data = example_summary,
  spatial_list = example_spatial,
  patient_id = "deidentified_id",
  sample_id = "deidentified_sample"
)

mnames_bad <- c("cd3_cd8", "cd3_foxp3", "cd3_opal_570_positive",
"cd8_opal_520_positive", "foxp3_opal_620_positive",
"pdl1_opal_540_positive", "pd1_opal_650_positive")

mif_tumor = subset_mif(mif = mif, classifier = 'Classifier.Label', 
level = 'Tumor', markers = markers)
}
