% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetLabels.R
\name{subsetLabels}
\alias{subsetLabels}
\title{subsetLabels}
\usage{
subsetLabels(
  file = "",
  cluster = NULL,
  export.all = TRUE,
  slice.n = "slice1",
  dir.out = ""
)
}
\arguments{
\item{file}{Seurat object as Input}

\item{cluster}{Cluster names in Idents}

\item{export.all}{Export coordinates files}

\item{slice.n}{Image slice number}

\item{dir.out}{directory output}
}
\description{
subsetLabels
}
\details{
This function allows the selection of clusters of interest and exporting their coordinates for further analysis
}
